Resolved issues commented on the PR
