+++ app/build.gradle
@@ -24,8 +24,6 @@ dependencies {
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-
-
     compile 'com.android.support:appcompat-v7:25.3.1'
     compile 'com.google.dagger:dagger:2.10'
     compile 'javax.inject:javax.inject:1'
