+++ app/src/main/AndroidManifest.xml
@@ -3,20 +3,20 @@
     package="com.globant.practice">
 
     <application
-        android:name=".Di.PracticeApplication"
+        android:name=".PracticeApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".View.Activities.SplashActivity">
+        <activity android:name=".view.activities.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".View.Activities.HomeActivity"></activity>
+        <activity android:name=".view.activities.HomeActivity"></activity>
     </application>
 
 </manifest>
