+++ app/src/main/java/com/globant/practice/di/PracticeModule.java
@@ -1,6 +1,6 @@
-package com.globant.practice.Di;
+package com.globant.practice.di;
 
-import com.globant.practice.Presenter.SplashPresenter;
+import com.globant.practice.presenter.SplashPresenter;
 
 import javax.inject.Singleton;
 
@@ -8,11 +8,15 @@ import dagger.Module;
 import dagger.Provides;
 
 /**
+ * Provide all the possible reference that Dagger can inject.
  * Created by jonathan.vargas on 30/03/2017.
  */
 @Module
 public class PracticeModule {
-
+    /**
+     * Returns a unique reference of SplashPresenter.
+     * @return SplashPresenter reference
+     */
     @Provides
     @Singleton
     SplashPresenter provideSplashPresenter() {
