--- app/src/main/res/values-en/strings.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="app_name">Practice</string>
-    <string name="splash_text">GitHub Suscriber App</string>
-</resources>
