+++ app/src/main/res/values-es/strings.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">Práctica</string>
     <string name="splash_text">Aplicación Suscriptores GitHub</string>
 </resources>
