+++ app/src/main/res/values/dimen.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="text_splash_size">34dp</dimen>
+</resources>
